refactor mixed quant and skip multimodal
