+++ mlx_vlm/__init__.py
@@ -1,8 +1,9 @@
 import os
 
+from .convert import convert
 from .generate import GenerationResult, generate, stream_generate
 from .prompt_utils import apply_chat_template, get_message_json
-from .utils import convert, load, prepare_inputs, process_image, quantize_model
+from .utils import load, prepare_inputs, process_image
 from .version import __version__
 
 os.environ["TRANSFORMERS_NO_ADVISORY_WARNINGS"] = "1"
