+++ mlx_vlm/convert.py
@@ -1,8 +1,180 @@
 import argparse
+import glob
+import shutil
+from pathlib import Path
+from typing import Callable, Optional, Union
 
-from .utils import MODEL_CONVERSION_DTYPES, convert
+import mlx.core as mx
+import mlx.nn as nn
+from mlx.utils import tree_map_with_path
+from mlx_lm.utils import dequantize_model, quantize_model, tree_flatten
+
+from .utils import (
+    MODEL_CONVERSION_DTYPES,
+    fetch_from_hub,
+    get_model_path,
+    save_config,
+    save_weights,
+    skip_multimodal_module,
+    upload_to_hub,
+)
 
 QUANT_RECIPES = ["mixed_2_6", "mixed_3_4", "mixed_3_6", "mixed_4_6"]
+
+
+def mixed_quant_predicate_builder(
+    recipe: str, model: nn.Module
+) -> Callable[[str, nn.Module, dict], Union[bool, dict]]:
+
+    high_bits = 6
+    group_size = 64
+
+    if recipe == "mixed_2_6":
+        low_bits = 2
+    elif recipe == "mixed_3_4":
+        low_bits = 3
+        high_bits = 4
+    elif recipe == "mixed_3_6":
+        low_bits = 3
+    elif recipe == "mixed_4_6":
+        low_bits = 4
+    else:
+        raise ValueError("Invalid quant recipe {recipe}")
+
+    down_keys = [k for k, _ in model.named_modules() if "down_proj" in k]
+    if len(down_keys) == 0:
+        raise ValueError("Model does not have expected keys for mixed quant.")
+
+    # Look for the layer index location in the path:
+    for layer_location, k in enumerate(down_keys[0].split(".")):
+        if k.isdigit():
+            break
+    num_layers = len(model.layers)
+
+    def mixed_quant_predicate(
+        path: str,
+        module: nn.Module,
+        config: dict,
+    ) -> Union[bool, dict]:
+        """Implements mixed quantization predicates with similar choices to, for example, llama.cpp's Q4_K_M.
+        Ref: https://github.com/ggerganov/llama.cpp/blob/917786f43d0f29b7c77a0c56767c0fa4df68b1c5/src/llama.cpp#L5265
+        By Alex Barron: https://gist.github.com/barronalex/84addb8078be21969f1690c1454855f3
+        """
+
+        if skip_multimodal_module(path):
+            return False
+        if not hasattr(module, "to_quantized"):
+            return False
+        if module.weight.shape[1] % group_size != 0:
+            return False
+
+        # Extract layer index from path, safely handling non-digit elements
+        path_parts = path.split(".")
+        index = 0  # Default to 0 for modules without layer indices
+        if len(path_parts) > layer_location:
+            element = path_parts[layer_location]
+            if element.isdigit():
+                index = int(element)
+        use_more_bits = (
+            index < num_layers // 8
+            or index >= 7 * num_layers // 8
+            or (index - num_layers // 8) % 3 == 2
+        )
+        if "v_proj" in path and use_more_bits:
+            return {"group_size": group_size, "bits": high_bits}
+        if "down_proj" in path and use_more_bits:
+            return {"group_size": group_size, "bits": high_bits}
+        if "lm_head" in path:
+            return {"group_size": group_size, "bits": high_bits}
+
+        return {"group_size": group_size, "bits": low_bits}
+
+    return mixed_quant_predicate
+
+
+def convert(
+    hf_path: str,
+    mlx_path: str = "mlx_model",
+    quantize: bool = False,
+    q_group_size: int = 64,
+    q_bits: int = 4,
+    dtype: Optional[str] = None,
+    upload_repo: str = None,
+    revision: Optional[str] = None,
+    dequantize: bool = False,
+    trust_remote_code: bool = True,
+    quant_predicate: Optional[str] = None,
+):
+    print("[INFO] Loading")
+    model_path = get_model_path(hf_path, revision=revision)
+    model, config, processor = fetch_from_hub(
+        model_path, lazy=True, trust_remote_code=trust_remote_code
+    )
+
+    def base_quant_predicate(path, module, config):
+        if skip_multimodal_module(path):
+            return False
+        if not hasattr(module, "to_quantized"):
+            return False
+        if module.weight.shape[1] % q_group_size != 0:
+            return False
+        return True
+
+    if isinstance(quant_predicate, str):
+        quant_predicate = mixed_quant_predicate_builder(quant_predicate, model)
+
+    quant_predicate = quant_predicate or base_quant_predicate
+
+    if dtype is None:
+        dtype = config.get("torch_dtype", None)
+    if dtype in MODEL_CONVERSION_DTYPES:
+        print("[INFO] Using dtype:", dtype)
+        dtype = getattr(mx, dtype)
+        cast_predicate = getattr(model, "cast_predicate", lambda _: True)
+
+        def set_dtype(k, v):
+            if cast_predicate(k) and mx.issubdtype(v.dtype, mx.floating):
+                return v.astype(dtype)
+            else:
+                return v
+
+        model.update(tree_map_with_path(set_dtype, model.parameters()))
+
+    if quantize and dequantize:
+        raise ValueError("Choose either quantize or dequantize, not both.")
+
+    if quantize:
+        print("[INFO] Quantizing")
+        config.setdefault("vision_config", {})
+        model, config = quantize_model(
+            model, config, q_group_size, q_bits, quant_predicate=quant_predicate
+        )
+
+        weights = dict(tree_flatten(model.parameters()))
+
+    if dequantize:
+        print("[INFO] Dequantizing")
+        model = dequantize_model(model)
+        weights = dict(tree_flatten(model.parameters()))
+
+    if isinstance(mlx_path, str):
+        mlx_path = Path(mlx_path)
+
+    del model
+    save_weights(mlx_path, weights, donate_weights=True)
+
+    # Copy Python and JSON files from the model path to the MLX path
+    for pattern in ["*.py", "*.json"]:
+        files = glob.glob(str(model_path / pattern))
+        for file in files:
+            shutil.copy(file, mlx_path)
+
+    processor.save_pretrained(mlx_path)
+
+    save_config(config, config_path=mlx_path / "config.json")
+
+    if upload_repo is not None:
+        upload_to_hub(mlx_path, upload_repo, hf_path)
 
 
 def configure_parser() -> argparse.ArgumentParser:
@@ -56,18 +228,6 @@ def configure_parser() -> argparse.ArgumentParser:
         action="store_true",
         default=False,
     )
-    parser.add_argument(
-        "--skip-vision",
-        help="Skip vision module quantization.",
-        action="store_true",
-        default=False,
-    )
-    parser.add_argument(
-        "--skip-audio",
-        help="Skip audio module quantization.",
-        action="store_true",
-        default=False,
-    )
     return parser
 
 
@@ -78,4 +238,8 @@ def main():
 
 
 if __name__ == "__main__":
+    print(
+        "Calling `python -m mlx_vlm.convert ...` directly is deprecated."
+        " Use `mlx_vlm.convert ...` or `python -m mlx_vlm convert ...` instead."
+    )
     main()
